requirements.txt
Django==1.11.2
django-leaflet==0.22.0
psycopg2==2.7.1
pytz==2017.2
